// src/app/models/fees.ts
export interface IFees {
  airtime: string;
  betting: string;
  cable: string;
  card_create_dollar: string;
  card_create_naira: string;
  card_fund_dollar: string;
  card_fund_naira: string;
  card_withdraw_naira: string;
  data: string;
  dollar_rate_charge: string;
  education: string;
  electricity: string;
  transfer_greater_than_limit: string;
  transfer_less_than_limit: string;
  transfer_limit_amount: string;
}